requirements.txt
pandas
numpy
requests
sqlalchemy
python-dotenv
pyarrow
pytest
matplotlib
seaborn
streamlit
psycopg2-binary
PyYAML
